examples_v2/requirements.txt
# assume torch is already installed

# pycolmap for data parsing
git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e
# nerfview for viewer
git+https://github.com/hangg7/nerfview@9dff4ef2dec489e17519f2124a92e023dea1254c
# (optional) nerfacc for torch version rasterization 
# git+https://github.com/nerfstudio-project/nerfacc

imageio
numpy
scikit-learn
tqdm
nerfview
torchmetrics[image]